% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlAGHQ.R
\name{controlAGHQ}
\alias{controlAGHQ}
\title{Set aghq::quad parameters for model fitting.}
\usage{
controlAGHQ(...)
}
\arguments{
\item{...}{See details.}
}
\value{
A list of parameters for use by aghq::aghq.
}
\description{
.
}
\details{
So far, controlAGHQ() holds only one parameter, `k`, the number of quadrature points.
}
\examples{
controlAGHQ()
controlAGHQ(k=5)
}
