% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitModel}
\alias{fitModel}
\title{Fit a model}
\usage{
fitModel(model, data, dll = NULL)
}
\arguments{
\item{model}{A model (list of info, see template).}

\item{data}{The dataset on which the model is fitted. It should contain all the variables involved in the model.}
}
\value{
A fitted model.
}
\description{
Function used for fitting the desired case crossover model.
}
