% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_specifiers.R
\name{fixedEffect}
\alias{fixedEffect}
\title{Specify a fixed effect.}
\usage{
fixedEffect(prior_prec = 0.01, as_factor = F)
}
\arguments{
\item{prior}{Prior distribution on the (beta) coefficient. Only the gaussian prior is implemented.}
}
\description{
Specify a fixed effect.
}
\examples{
fixedEffect()
}
