% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{generateSamples}
\alias{generateSamples}
\title{Generate samples from the posterior distribution.}
\usage{
generateSamples(fit, M = 10000)
}
\arguments{
\item{fit}{A fitted case-crossover model (obtained via `fitModel(model)`).}

\item{M}{Number of replicates to draw for the posterior (to approximate it, the more the better).}
}
\value{
A list of length three with data.frames containing samples from the posterior
distribution of beta, z and theta, respectively.
}
\description{
Generate samples from the posterior distribution.
}
