% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{getResults}
\alias{getResults}
\title{Extract results from a fitted ccModel.}
\usage{
getResults(fit, probs = 0.5, M = 10000)
}
\arguments{
\item{fit}{A fitted ccModel (obtained via `fitModel(model)`).}

\item{quantiles}{Quantiles of the posterior (cumulative) distribution of fixed and random effects (and overdispersion terms).}
}
\value{
A data.frame with values of the beta and gamma coefficients, as well as their credible intervals.
}
\description{
Extract results from a fitted ccModel.
}
