% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_specifiers.R
\name{pc_prec_prior}
\alias{pc_prec_prior}
\title{Specify an exponential (penalized complexity) prior distribution.}
\usage{
pc_prec_prior(alpha = 0.5, u = 0.001)
}
\arguments{
\item{alpha}{A probability.}

\item{u}{A value greater than 0.}
}
\value{
A list specifying a pc prep prior distribution.
}
\description{
PC prior distribution; see \insertCite{Simpson:2017;textual}{bayesEpi}. The parameters are such that \eqn{P(\sigma > u) = \alpha}, where \eqn{\sigma = -2log(\theta)}.
}
\examples{
pc_prec_prior(alpha = .5, u = .01)
}
\references{
{
  \insertRef{Simpson:2017}{bayesEpi}
}
}
